Move all query modifications to a separate file
